# .mem/openapi.yaml
openapi: 3.0.3
info:
  title: Co-Work API
  description: API documentation for the Co-Work collaborative project management system
  version: 1.0.0
  contact:
    name: Co-Work Team
servers:
  - url: /api
    description: API base path
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
...
    File:
      type: object
      properties:
        id:
          type: integer
          description: File ID
        projectId:
          type: integer
          description: Project ID
        parentId:
          type: integer
          nullable: true
          description: Parent folder ID
        name:
          type: string
          description: File or folder name
        type:
          type: string
          enum: [file, folder]
          description: Resource type
        fileType:
          type: string
          enum: [docx, md, txt, code, image]
          nullable: true
          description: File type
        content:
          type: string
          nullable: true
          description: File content
        ownerType:
          type: string
          enum: [team, private]
          description: Owner type
        ownerId:
          type: integer
          nullable: true
          description: Owner user ID
        status:
          type: string
          enum: [modified, new, synced]
          description: File status
        modifiedAt:
          type: string
          format: date-time
          description: Last modification timestamp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - projectId
        - name
        - type
        - ownerType
        - status
        - modifiedAt
        - createdAt
...
paths:
...
  /files:
    get:
      tags:
        - Files
      summary: List files
      description: Retrieve files and folders with filtering options
      parameters:
        - name: id
          in: query
          description: File ID for single file retrieval
          schema:
            type: integer
        - name: projectId
          in: query
          description: Project ID (required for listing files)
          schema:
            type: integer
        - name: ownerType
          in: query
          description: Filter by owner type
          schema:
            type: string
            enum: [team, private]
        - name: userId
          in: query
          description: Filter by owner user ID
          schema:
            type: integer
        - name: parentId
          in: query
          description: Filter by parent folder ID (use 'null' or empty for root)
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results (max 100)
          schema:
            type: integer
            default: 100
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/File'
                  - type: array
                    items:
                      $ref: '#/components/schemas/File'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Files
      summary: Create file or folder
      description: Create a new file or folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - name
                - type
                - ownerType
              properties:
                projectId:
                  type: integer
                  description: Project ID
                name:
                  type: string
                  description: File or folder name
                type:
                  type: string
                  enum: [file, folder]
                  description: Resource type
                ownerType:
                  type: string
                  enum: [team, private]
                  description: Owner type
                parentId:
                  type: integer
                  nullable: true
                  description: Parent folder ID
                fileType:
                  type: string
                  enum: [docx, md, txt, code, image]
                  description: File type
                content:
                  type: string
                  description: File content
                ownerId:
                  type: integer
                  description: Owner user ID
                status:
                  type: string
                  enum: [modified, new, synced]
                  description: File status
      responses:
        '201':
          description: File created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Files
      summary: Update file
      description: Update an existing file or folder
      parameters:
        - name: id
          in: query
          required: true
          description: File ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: File or folder name
                content:
                  type: string
                  description: File content
                status:
                  type: string
                  enum: [modified, new, synced]
                  description: File status
                parentId:
                  type: integer
                  nullable: true
                  description: Parent folder ID
                fileType:
                  type: string
                  enum: [docx, md, txt, code, image]
                  description: File type
      responses:
        '200':
          description: File updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Files
      summary: Delete file
      description: Delete an existing file or folder
      parameters:
        - name: id
          in: query
          required: true
          description: File ID
          schema:
            type: integer
      responses:
        '200':
          description: File deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  file:
                    $ref: '#/components/schemas/File'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...